requirements.txt
biopython==1.72
click==7.0
flask-bootstrap==3.3.7.1
flask-mail==0.9.1
flask-wtf==0.14.2
flask==1.0.2
gunicorn
pandas==0.25.3
python-dotenv==0.9.1
requests==2.20.1
werkzeug==0.16.1
sklearn